Update scripts
Add validation to avoid running the pgdump restore again
Update claims Hierarchy insert to update or insert the value

+++ eng/docker-compose/setup-database-template.psm1
@@ -39,16 +39,22 @@ function LoadSeedData {
 
     if ($LASTEXITCODE -eq 0) {
         Write-Host "PostgreSQL is running. Proceeding with bootstrap..."
+        $schemaName = "dms"
+        $schemaExists = docker exec dms-postgresql psql -U $User -d $Database -tAc "SELECT 1 FROM pg_namespace WHERE nspname = '$schemaName';"
+        if (-not $schemaExists) {
+            docker cp $sqlPath "dms-postgresql:/tmp/restore.sql"
+            docker exec dms-postgresql psql -U postgres -d $database -f /tmp/restore.sql
 
-        docker cp $sqlPath "dms-postgresql:/tmp/restore.sql"
-        docker exec dms-postgresql psql -U postgres -d $database -f /tmp/restore.sql
-
-        if ($LASTEXITCODE -eq 0) {
-            Write-Host "Template data loaded successfully."
-        } else {
+            if ($LASTEXITCODE -eq 0) {
+                Write-Host "Template data loaded successfully."
+            } else {
+                Write-Error "Failed to execute the database script file."
+            }
+        }
+        else {
             Write-Error "Failed to execute the database script file."
         }
     } else {
-        Write-Error "PostgreSQL server is not running or unreachable."
+        Write-Error "PostgreSQL Data Seed Load failed: existing volumes were detected. Please remove them before proceeding."
     }
 }
